Add testcafe as explicit dev dep
- We need types to do proper linting and type checking
- Include types for DX
- Makes the e2e tests run a lot faster since we skip the npx step
- Allow a testcafe version update to be verified before being forced

Since both Jest and TestCafe exports global type definitions
for common names, we cannot use the same ts project for it.
Having a separate tsconfig.json in the e2e directory so
that Jest and TestCafe files are imported in each respective
tsconfig.json, and specifying it in eslint config, seems to
solve this issue.

This also fixes so that eslist is run on all files which
are not specified in .eslintignore.